Refactor layout and add not found page
